refactor(rules)!: weights become integer centinats, and ties become visible
Rule weights were `f32` nats.  They are now `i16` hundredths of a nat —
`1.55` is `155` — across `Rules::rule`, `Rule::weight`, `row`, and
`set_natural_double_weight`.

No board moved.  `smoke-default` over 20 000 seeded boards is
byte-identical, and `nats(c) = c as f32 / 100.0` is bit-for-bit the old
decimal literal for all 75 distinct authored weights: `i16 -> f32` is
exact and IEEE division is correctly rounded.  That second proof covers
`dutch()` and `instinct.rs`, which the board diff does not.  The lone
exception is an authored `-0.0`, inert because `eval` adds the weight to
a crisp `0.0`/`-inf` and `-0.0 + 0.0` is `+0.0`.

The unit buys an equality `f32` could not give.  Four tables built a
declining ladder by repeated subtraction, and the drift was silent and
site-dependent: the 2/1 up-the-line ladder in `responses.rs` reached
`1.000000119` where it meant `1.0` — comparing *unequal* to a rule
authored `1.0` — while the same idiom in `notrump.rs` stayed exact.
Integer rungs make "these two rules share a weight" decidable, which the
row layer needs before generic auction-pattern rows can detect their own
collisions.

On that footing, `rows::weight_tie_report` (asserted by
`assert_package_invariants`) reports every rung claimed twice in one
table, naming each rule's authoring site via `#[track_caller]` on
`Rules::rule`, exposed as `Rule::origin`.  Such a pair is redundant — the
logit is the max over crisp constraints, so it is exactly
`w + crisp(C1 or C2)`, and the reader already disjoins every matching
rule's projection — and lossy, since `Rules::explain` breaks the tie with
a strict `>` and names only the first rule, leaving authoring order to
decide which alert and label describe the call.

The census found two standing clusters, both in `direct-seat` and both
one convention partitioned across the auction cases its row pattern
spans, rather than a precedence bug: the Modern negative double (four
arms on `min_level_is`/`they_bid`, HCP floors 6/6/8/8) and the
invitational `2NT` (two arms with the same hand gate, split only on
cheapest-versus-jump).  Both are recorded in `KNOWN_WEIGHT_TIES`;
retiring either is a bidding change and wants the A/B first.

`examples/dump-corpus` still emits `weight` in nats, converted at the
boundary, so its JSON consumers see unchanged numbers.

+++ examples/bba-gen/main.rs
@@ -891,10 +891,10 @@ struct Args {
     #[arg(long, default_value = "8:14")]
     ns_overcall: String,
 
-    /// Logit weight of our natural penalty double of their 1NT (default 1.3, above
-    /// the 1.0 suit overcall).
-    #[arg(long, default_value_t = 1.3)]
-    ns_double_weight: f32,
+    /// Logit weight of our natural penalty double of their 1NT, in centinats
+    /// (default 130, above the 100 suit overcall).
+    #[arg(long, default_value_t = 130)]
+    ns_double_weight: i16,
 
     /// Support gate on our 12+ takeout double of a suit / weak-two opening:
     /// off | lenient | strict (default, matches shipped `american()`).
